requirements.txt
pystray>=0.15
Pillow>=9.0.0
